requirements.txt
fastapi
uvicorn
boto3
python-jose[cryptography]
python-multipart
httpx
pillow
tenacity
python-dotenv
requests